pyproject.toml
[tool]

[tool.poetry]
name = "boxsup_pytorch"
version = "0.0.0"
homepage = "https://github.com/makanu/boxsup-pytorch"
description = "The BoxSup-pytorch is an implementation of the BoxSup Algorithm in pytorch. Since the original paper has no available implementation this implementation is only based on the paper. They may be a few differences to the explanations from the paper."
authors = ["Matti Kaupenjohann <matti.kaupenjohann@fh-dortmund.de>"]
readme = "README.md"
license =  "GPLv3"
classifiers=[
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
]
include = [
    "LICENSE",
    "boxsup_pytorch/py.typed"
]


[tool.poetry.dependencies]
python = ">=3.8,<3.11"
importlib_metadata = ">=4.4.0"
numpy = "1.22.3"
Pillow = ">=9.1.0, <10.0.0"
ninja = "^1.10.2"
tqdm = "^4.64.0"
scipy = "^1.8.0"
torch = [
  {url = "https://download.pytorch.org/whl/cu116/torch-1.12.1%2Bcu116-cp310-cp310-linux_x86_64.whl", markers = "sys_platform == 'linux'"},
  {url = "https://download.pytorch.org/whl/cu116/torch-1.12.1%2Bcu116-cp310-cp310-win_amd64.whl", markers = "sys_platform == 'win32'"}
]
torchvision = "^0.12.0"
configparser = "^5.2.0"
alive-progress = "^2.4.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.8.0"


[tool.poetry.dev-dependencies]
mkdocs-material = "^6.1.5"
pytest = "^6.0"
pytest-cov = "^2.10.1"
flake8 = "^3.8.4"
flake8-docstrings = "1.5.0"
flake8-import-order = "0.18.1"
flake8-colors = "0.1.6"

black = "22.3.0"
mypy = "0.950"
yamllint = "^1.25.0"
pep440 = "^0.1.0"
[tool.black]
line-length=100

[tool.pylint.reports]
output-format="colorized"
reports="y"
include-ids="yes"
msg-template="{msg_id}:{line:3d},{column}: {obj}: {msg}"

[tool.pytest.ini_options]
addopts = "--cov-config=.coveragerc --cov=boxsup_pytorch --cov-report term-missing  -vv --color=yes --cov-fail-under 80"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
